Run tests from development version of app.
