+++ public/components/log.js
@@ -1,6 +1,6 @@
 const log = require('electron-log');
 
-if (process.env.NODE_ENV !== 'development') {
+if (process.env.NODE_ENV !== 'development' || process.env.NODE_ENV !== 'test') {
   log.transports.console.level = false;
 }
 log.transports.file.level = 'log';
