+++ test/README.md
@@ -1,0 +1,15 @@
+# End to end tests
+
+Uses spectron to run app in development mode.
+
+Because of older version of Electron, version 4.0.0 is used, which correlates to
+version 4 of webdriver, docs which can be found here:
+
+http://v4.webdriver.io/api.html
+
+# Run tests:
+
+```
+$ npm run start:electron
+$ npm run test:e2e
+```
