+++ test/test/helpers.js
@@ -2,17 +2,14 @@ import path from 'path';
 import { Application } from 'spectron';
 import electron from 'electron';
 
-// Each app may depend on different electron versions, so we require from
-// their respective node_modules.
-// `require('electron')` returns the path to the binary.
-// eslint-disable-next-line global-require, import/no-dynamic-require
-// const getElectronBinaryPath = () => require(path.join(__dirname, 'node_modules', 'electron'));
-
 export const makeTestingApp = () => {
-  // const electronPath = getElectronBinaryPath(appName);
-  const appBuild = path.join(__dirname, '../', '../', 'www');
+  const appBuild = path.join(__dirname, '../', '../', 'public');
   const appConfig = {
     path: electron,
+    env: {
+      NODE_ENV: 'test',
+      NC_DEVSERVER_FILE: '.devserver',
+    },
     args: [appBuild],
   };
   return new Application(appConfig);
@@ -20,7 +17,7 @@ export const makeTestingApp = () => {
 
 export const startApp = async (app) => {
   await app.start();
-  await app.browserWindow.isVisible();
+  await app.client.waitUntilWindowLoaded();
   return app;
 };
